delete ResultBox, minor text adjust learner table reuslt
